@@ -37,6 +37,10 @@ var defaultModelRatio = map[string]float64{
 	"gpt-4o":                         2.5,  // $0.005 / 1K tokens
 	"gpt-4o-2024-05-13":              2.5,  // $0.005 / 1K tokens
 	"gpt-4o-2024-08-06":              1.25, // $0.01 / 1K tokens
+	"o1-preview":                     7.5,
+	"o1-preview-2024-09-12":          7.5,
+	"o1-mini":                        1.5,
+	"o1-mini-2024-09-12":             1.5,
 	"gpt-4-turbo":                    5,    // $0.01 / 1K tokens
 	"gpt-4-turbo-2024-04-09":         5,    // $0.01 / 1K tokens
 	"gpt-4-1106-preview":             5,    // $0.01 / 1K tokens
@@ -81,6 +85,7 @@ var defaultModelRatio = map[string]float64{
 	"claude-2.1":                     4,     // $8 / 1M tokens
 	"claude-3-haiku-20240307":        0.125, // $0.25 / 1M tokens
 	"claude-3-5-sonnet-20240620":     1.5,   // $3 / 1M tokens
+	"claude-3-5-sonnet-20241022":     1.5,   // $3 / 1M tokens
 	"claude-3-sonnet-20240229":       1.5,   // $3 / 1M tokens
 	"claude-3-opus-20240229":         7.5,   // $15 / 1M tokens
 	"ERNIE-4.0-8K":                   0.120 * RMB,
@@ -104,8 +109,10 @@ var defaultModelRatio = map[string]float64{
 	"gemini-pro-vision":              1, // $0.00025 / 1k characters -> $0.001 / 1k tokens
 	"gemini-1.0-pro-vision-001":      1,
 	"gemini-1.0-pro-001":             1,
-	"gemini-1.5-pro-latest":          1,
+	"gemini-1.5-pro-latest":          1.75, // $3.5 / 1M tokens
+	"gemini-1.5-pro-exp-0827":        1.75, // $3.5 / 1M tokens
 	"gemini-1.5-flash-latest":        1,
+	"gemini-1.5-flash-exp-0827":      1,
 	"gemini-1.0-pro-latest":          1,
 	"gemini-1.0-pro-vision-latest":   1,
 	"gemini-ultra":                   1,
@@ -117,6 +124,13 @@ var defaultModelRatio = map[string]float64{
 	"glm-4v":                         0.05 * RMB, // ￥0.05 / 1k tokens
 	"glm-4-alltools":                 0.1 * RMB,  // ￥0.1 / 1k tokens
 	"glm-3-turbo":                    0.3572,
+	"glm-4-plus":                     0.05 * RMB,
+	"glm-4-0520":                     0.1 * RMB,
+	"glm-4-air":                      0.001 * RMB,
+	"glm-4-airx":                     0.01 * RMB,
+	"glm-4-long":                     0.001 * RMB,
+	"glm-4-flash":                    0,
+	"glm-4v-plus":                    0.01 * RMB,
 	"qwen-turbo":                     0.8572, // ￥0.012 / 1k tokens
 	"qwen-plus":                      10,     // ￥0.14 / 1k tokens
 	"text-embedding-v1":              0.05,   // ￥0.0007 / 1k tokens
@@ -135,26 +149,28 @@ var defaultModelRatio = map[string]float64{
 	"hunyuan":                        7.143,  // ¥0.1 / 1k tokens  // https://cloud.tencent.com/document/product/1729/97731#e0e6be58-60c8-469f-bdeb-6c264ce3b4d0
 	// https://platform.lingyiwanwu.com/docs#-计费单元
 	// 已经按照 7.2 来换算美元价格
-	"yi-34b-chat-0205":      0.18,
-	"yi-34b-chat-200k":      0.864,
-	"yi-vl-plus":            0.432,
-	"yi-large":              20.0 / 1000 * RMB,
-	"yi-medium":             2.5 / 1000 * RMB,
-	"yi-vision":             6.0 / 1000 * RMB,
-	"yi-medium-200k":        12.0 / 1000 * RMB,
-	"yi-spark":              1.0 / 1000 * RMB,
-	"yi-large-rag":          25.0 / 1000 * RMB,
-	"yi-large-turbo":        12.0 / 1000 * RMB,
-	"yi-large-preview":      20.0 / 1000 * RMB,
-	"yi-large-rag-preview":  25.0 / 1000 * RMB,
-	"command":               0.5,
-	"command-nightly":       0.5,
-	"command-light":         0.5,
-	"command-light-nightly": 0.5,
-	"command-r":             0.25,
-	"command-r-plus	":       1.5,
-	"deepseek-chat":         0.07,
-	"deepseek-coder":        0.07,
+	"yi-34b-chat-0205":       0.18,
+	"yi-34b-chat-200k":       0.864,
+	"yi-vl-plus":             0.432,
+	"yi-large":               20.0 / 1000 * RMB,
+	"yi-medium":              2.5 / 1000 * RMB,
+	"yi-vision":              6.0 / 1000 * RMB,
+	"yi-medium-200k":         12.0 / 1000 * RMB,
+	"yi-spark":               1.0 / 1000 * RMB,
+	"yi-large-rag":           25.0 / 1000 * RMB,
+	"yi-large-turbo":         12.0 / 1000 * RMB,
+	"yi-large-preview":       20.0 / 1000 * RMB,
+	"yi-large-rag-preview":   25.0 / 1000 * RMB,
+	"command":                0.5,
+	"command-nightly":        0.5,
+	"command-light":          0.5,
+	"command-light-nightly":  0.5,
+	"command-r":              0.25,
+	"command-r-plus":         1.5,
+	"command-r-08-2024":      0.075,
+	"command-r-plus-08-2024": 1.25,
+	"deepseek-chat":          0.07,
+	"deepseek-coder":         0.07,
 	// Perplexity online 模型对搜索额外收费，有需要应自行调整，此处不计入搜索费用
 	"llama-3-sonar-small-32k-chat":   0.2 / 1000 * USD,
 	"llama-3-sonar-small-32k-online": 0.2 / 1000 * USD,
@@ -335,14 +351,18 @@ func GetCompletionRatio(name string) float64 {
 		return 4.0 / 3.0
 	}
 	if strings.HasPrefix(name, "gpt-4") && name != "gpt-4-all" && name != "gpt-4-gizmo-*" {
-		if strings.HasPrefix(name, "gpt-4o-mini") || "gpt-4o-2024-08-06" == name {
+		if strings.HasSuffix(name, "preview") || strings.HasPrefix(name, "gpt-4-turbo") || "gpt-4o-2024-05-13" == name {
+			return 3
+		}
+
+		if strings.HasPrefix(name, "gpt-4o") {
 			return 4
 		}
 
-		if strings.HasSuffix(name, "preview") || strings.HasPrefix(name, "gpt-4-turbo") || strings.HasPrefix(name, "gpt-4o") {
-			return 3
-		}
 		return 2
+	}
+	if "o1" == name || strings.HasPrefix(name, "o1-") {
+		return 4
 	}
 	if name == "chatgpt-4o-latest" {
 		return 3
@@ -358,7 +378,7 @@ func GetCompletionRatio(name string) float64 {
 		return 3
 	}
 	if strings.HasPrefix(name, "gemini-") {
-		return 3
+		return 4
 	}
 	if strings.HasPrefix(name, "command") {
 		switch name {
@@ -366,6 +386,10 @@ func GetCompletionRatio(name string) float64 {
 			return 3
 		case "command-r-plus":
 			return 5
+		case "command-r-08-2024":
+			return 4
+		case "command-r-plus-08-2024":
+			return 4
 		default:
 			return 2
 		}
