@@ -20,14 +20,16 @@ var GetMediaTokenNotStream = common.GetEnvOrDefaultBool("GET_MEDIA_TOKEN_NOT_STR
 var UpdateTask = common.GetEnvOrDefaultBool("UPDATE_TASK", true)
 
 var GeminiModelMap = map[string]string{
-	"gemini-1.5-pro-latest":   "v1beta",
-	"gemini-1.5-pro-001":      "v1beta",
-	"gemini-1.5-pro":          "v1beta",
-	"gemini-1.5-pro-exp-0801": "v1beta",
-	"gemini-1.5-flash-latest": "v1beta",
-	"gemini-1.5-flash-001":    "v1beta",
-	"gemini-1.5-flash":        "v1beta",
-	"gemini-ultra":            "v1beta",
+	"gemini-1.5-pro-latest":     "v1beta",
+	"gemini-1.5-pro-001":        "v1beta",
+	"gemini-1.5-pro":            "v1beta",
+	"gemini-1.5-pro-exp-0801":   "v1beta",
+	"gemini-1.5-pro-exp-0827":   "v1beta",
+	"gemini-1.5-flash-latest":   "v1beta",
+	"gemini-1.5-flash-exp-0827": "v1beta",
+	"gemini-1.5-flash-001":      "v1beta",
+	"gemini-1.5-flash":          "v1beta",
+	"gemini-ultra":              "v1beta",
 }
 
 func InitEnv() {
@@ -44,3 +46,6 @@ func InitEnv() {
 		}
 	}
 }
+
+// 是否生成初始令牌，默认关闭。
+var GenerateDefaultToken = common.GetEnvOrDefaultBool("GENERATE_DEFAULT_TOKEN", false)
