@@ -4,8 +4,8 @@ import "C"
 import (
 	"fmt"
 	"github.com/gin-gonic/gin"
-	"github.com/stripe/stripe-go/v76"
-	"github.com/stripe/stripe-go/v76/checkout/session"
+	"github.com/stripe/stripe-go/v79"
+	"github.com/stripe/stripe-go/v79/checkout/session"
 	"log"
 	"one-api/common"
 	"one-api/model"
