@@ -1,6 +1,8 @@
 package aws
 
-import "one-api/relay/channel/claude"
+import (
+	"one-api/relay/channel/claude"
+)
 
 type AwsClaudeRequest struct {
 	// AnthropicVersion should be "bedrock-2023-05-31"
@@ -12,4 +14,6 @@ type AwsClaudeRequest struct {
 	TopP             float64                `json:"top_p,omitempty"`
 	TopK             int                    `json:"top_k,omitempty"`
 	StopSequences    []string               `json:"stop_sequences,omitempty"`
+	Tools            []claude.Tool          `json:"tools,omitempty"`
+	ToolChoice       any                    `json:"tool_choice,omitempty"`
 }
