@@ -1,7 +1,10 @@
 package cohere
 
 var ModelList = []string{
-	"command-r", "command-r-plus", "command-light", "command-light-nightly", "command", "command-nightly",
+	"command-r", "command-r-plus",
+	"command-r-08-2024", "command-r-plus-08-2024",
+	"c4ai-aya-23-35b", "c4ai-aya-23-8b",
+	"command-light", "command-light-nightly", "command", "command-nightly",
 	"rerank-english-v3.0", "rerank-multilingual-v3.0", "rerank-english-v2.0", "rerank-multilingual-v2.0",
 }
 
