@@ -21,11 +21,12 @@ import Redemption from './pages/Redemption';
 import TopUp from './pages/TopUp';
 import Log from './pages/Log';
 import Chat from './pages/Chat';
+import Chat2Link from './pages/Chat2Link';
 import { Layout } from '@douyinfe/semi-ui';
 import Midjourney from './pages/Midjourney';
 import Pricing from './pages/Pricing/index.js';
 import Task from './pages/Task/index.js';
-// import Detail from './pages/Detail';
+import Playground from './components/Playground.js';
 
 const Home = lazy(() => import('./pages/Home'));
 const Detail = lazy(() => import('./pages/Detail'));
@@ -59,215 +60,232 @@ function App() {
   }, []);
 
   return (
-    <Layout>
-      <Layout.Content>
-        <Routes>
-          <Route
-            path='/'
-            element={
+    <>
+      <Routes>
+        <Route
+          path='/'
+          element={
+            <Suspense fallback={<Loading></Loading>}>
+              <Home />
+            </Suspense>
+          }
+        />
+        <Route
+          path='/channel'
+          element={
+            <PrivateRoute>
+              <Channel />
+            </PrivateRoute>
+          }
+        />
+        <Route
+          path='/channel/edit/:id'
+          element={
+            <Suspense fallback={<Loading></Loading>}>
+              <EditChannel />
+            </Suspense>
+          }
+        />
+        <Route
+          path='/channel/add'
+          element={
+            <Suspense fallback={<Loading></Loading>}>
+              <EditChannel />
+            </Suspense>
+          }
+        />
+        <Route
+          path='/token'
+          element={
+            <PrivateRoute>
+              <Token />
+            </PrivateRoute>
+          }
+        />
+        <Route
+          path='/playground'
+          element={
+            <PrivateRoute>
+              <Playground />
+            </PrivateRoute>
+          }
+        />
+        <Route
+          path='/redemption'
+          element={
+            <PrivateRoute>
+              <Redemption />
+            </PrivateRoute>
+          }
+        />
+        <Route
+          path='/user'
+          element={
+            <PrivateRoute>
+              <User />
+            </PrivateRoute>
+          }
+        />
+        <Route
+          path='/user/edit/:id'
+          element={
+            <Suspense fallback={<Loading></Loading>}>
+              <EditUser />
+            </Suspense>
+          }
+        />
+        <Route
+          path='/user/edit'
+          element={
+            <Suspense fallback={<Loading></Loading>}>
+              <EditUser />
+            </Suspense>
+          }
+        />
+        <Route
+          path='/user/reset'
+          element={
+            <Suspense fallback={<Loading></Loading>}>
+              <PasswordResetConfirm />
+            </Suspense>
+          }
+        />
+        <Route
+          path='/login'
+          element={
+            <Suspense fallback={<Loading></Loading>}>
+              <LoginForm />
+            </Suspense>
+          }
+        />
+        <Route
+          path='/register'
+          element={
+            <Suspense fallback={<Loading></Loading>}>
+              <RegisterForm />
+            </Suspense>
+          }
+        />
+        <Route
+          path='/reset'
+          element={
+            <Suspense fallback={<Loading></Loading>}>
+              <PasswordResetForm />
+            </Suspense>
+          }
+        />
+        <Route
+          path='/oauth/github'
+          element={
+            <Suspense fallback={<Loading></Loading>}>
+              <GitHubOAuth />
+            </Suspense>
+          }
+        />
+        <Route
+          path='/oauth/linuxdo'
+          element={
+            <Suspense fallback={<Loading></Loading>}>
+              <LinuxDoOAuth />
+            </Suspense>
+          }
+        />
+        <Route
+          path='/setting'
+          element={
+            <PrivateRoute>
               <Suspense fallback={<Loading></Loading>}>
-                <Home />
+                <Setting />
               </Suspense>
-            }
-          />
-          <Route
-            path='/channel'
-            element={
-              <PrivateRoute>
-                <Channel />
-              </PrivateRoute>
-            }
-          />
-          <Route
-            path='/channel/edit/:id'
-            element={
+            </PrivateRoute>
+          }
+        />
+        <Route
+          path='/topup'
+          element={
+            <PrivateRoute>
               <Suspense fallback={<Loading></Loading>}>
-                <EditChannel />
+                <TopUp />
               </Suspense>
-            }
-          />
-          <Route
-            path='/channel/add'
-            element={
+            </PrivateRoute>
+          }
+        />
+        <Route
+          path='/log'
+          element={
+            <PrivateRoute>
+              <Log />
+            </PrivateRoute>
+          }
+        />
+        <Route
+          path='/detail'
+          element={
+            <PrivateRoute>
               <Suspense fallback={<Loading></Loading>}>
-                <EditChannel />
+                <Detail />
               </Suspense>
-            }
-          />
-          <Route
-            path='/token'
-            element={
-              <PrivateRoute>
-                <Token />
-              </PrivateRoute>
-            }
-          />
-          <Route
-            path='/redemption'
-            element={
-              <PrivateRoute>
-                <Redemption />
-              </PrivateRoute>
-            }
-          />
-          <Route
-            path='/user'
-            element={
-              <PrivateRoute>
-                <User />
-              </PrivateRoute>
-            }
-          />
-          <Route
-            path='/user/edit/:id'
-            element={
+            </PrivateRoute>
+          }
+        />
+        <Route
+          path='/midjourney'
+          element={
+            <PrivateRoute>
               <Suspense fallback={<Loading></Loading>}>
-                <EditUser />
+                <Midjourney />
               </Suspense>
-            }
-          />
-          <Route
-            path='/user/edit'
-            element={
+            </PrivateRoute>
+          }
+        />
+        <Route
+          path='/task'
+          element={
+            <PrivateRoute>
               <Suspense fallback={<Loading></Loading>}>
-                <EditUser />
+                <Task />
               </Suspense>
-            }
-          />
-          <Route
-            path='/user/reset'
-            element={
+            </PrivateRoute>
+          }
+        />
+        <Route
+          path='/pricing'
+          element={
+            <Suspense fallback={<Loading></Loading>}>
+              <Pricing />
+            </Suspense>
+          }
+        />
+        <Route
+          path='/about'
+          element={
+            <Suspense fallback={<Loading></Loading>}>
+              <About />
+            </Suspense>
+          }
+        />
+        <Route
+          path='/chat/:id?'
+          element={
+            <Suspense fallback={<Loading></Loading>}>
+              <Chat />
+            </Suspense>
+          }
+        />
+        {/* 方便使用chat2link直接跳转聊天... */}
+        <Route
+          path='/chat2link'
+          element={
+            <PrivateRoute>
               <Suspense fallback={<Loading></Loading>}>
-                <PasswordResetConfirm />
+                <Chat2Link />
               </Suspense>
-            }
-          />
-          <Route
-            path='/login'
-            element={
-              <Suspense fallback={<Loading></Loading>}>
-                <LoginForm />
-              </Suspense>
-            }
-          />
-          <Route
-            path='/register'
-            element={
-              <Suspense fallback={<Loading></Loading>}>
-                <RegisterForm />
-              </Suspense>
-            }
-          />
-          <Route
-            path='/reset'
-            element={
-              <Suspense fallback={<Loading></Loading>}>
-                <PasswordResetForm />
-              </Suspense>
-            }
-          />
-          <Route
-            path='/oauth/github'
-            element={
-              <Suspense fallback={<Loading></Loading>}>
-                <GitHubOAuth />
-              </Suspense>
-            }
-          />
-          <Route
-            path='/oauth/linuxdo'
-            element={
-              <Suspense fallback={<Loading></Loading>}>
-                <LinuxDoOAuth />
-              </Suspense>
-            }
-          />
-          <Route
-            path='/setting'
-            element={
-              <PrivateRoute>
-                <Suspense fallback={<Loading></Loading>}>
-                  <Setting />
-                </Suspense>
-              </PrivateRoute>
-            }
-          />
-          <Route
-            path='/topup'
-            element={
-              <PrivateRoute>
-                <Suspense fallback={<Loading></Loading>}>
-                  <TopUp />
-                </Suspense>
-              </PrivateRoute>
-            }
-          />
-          <Route
-            path='/log'
-            element={
-              <PrivateRoute>
-                <Log />
-              </PrivateRoute>
-            }
-          />
-          <Route
-            path='/detail'
-            element={
-              <PrivateRoute>
-                <Suspense fallback={<Loading></Loading>}>
-                  <Detail />
-                </Suspense>
-              </PrivateRoute>
-            }
-          />
-          <Route
-            path='/midjourney'
-            element={
-              <PrivateRoute>
-                <Suspense fallback={<Loading></Loading>}>
-                  <Midjourney />
-                </Suspense>
-              </PrivateRoute>
-            }
-          />
-          <Route
-            path='/task'
-            element={
-              <PrivateRoute>
-                <Suspense fallback={<Loading></Loading>}>
-                  <Task />
-                </Suspense>
-              </PrivateRoute>
-            }
-          />
-          <Route
-            path='/pricing'
-            element={
-              <Suspense fallback={<Loading></Loading>}>
-                <Pricing />
-              </Suspense>
-            }
-          />
-          <Route
-            path='/about'
-            element={
-              <Suspense fallback={<Loading></Loading>}>
-                <About />
-              </Suspense>
-            }
-          />
-          <Route
-            path='/chat'
-            element={
-              <Suspense fallback={<Loading></Loading>}>
-                <Chat />
-              </Suspense>
-            }
-          />
-          <Route path='*' element={<NotFound />} />
-        </Routes>
-      </Layout.Content>
-    </Layout>
+            </PrivateRoute>
+          }
+        />
+        <Route path='*' element={<NotFound />} />
+      </Routes>
+    </>
   );
 }
 
