@@ -3,7 +3,6 @@ import ReactDOM from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
 import App from './App';
 import HeaderBar from './components/HeaderBar';
-import Footer from './components/Footer';
 import 'semantic-ui-offline/semantic.min.css';
 import './index.css';
 import { UserProvider } from './context/User';
@@ -13,35 +12,40 @@ import { StatusProvider } from './context/Status';
 import { Layout } from '@douyinfe/semi-ui';
 import SiderBar from './components/SiderBar';
 import { ThemeProvider } from './context/Theme';
+import FooterBar from './components/Footer';
 
 // initialization
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
-const { Sider, Content, Header } = Layout;
+const { Sider, Content, Header, Footer } = Layout;
 root.render(
   <React.StrictMode>
     <StatusProvider>
       <UserProvider>
         <BrowserRouter>
           <ThemeProvider>
-            <Layout>
-              <Sider>
-                <SiderBar />
-              </Sider>
-              <Layout>
-                <Header>
-                  <HeaderBar />
-                </Header>
-                <Content
-                  style={{
-                    padding: '24px',
-                  }}
-                >
-                  <App />
-                </Content>
-                <Layout.Footer>
-                  <Footer></Footer>
-                </Layout.Footer>
+            <Layout
+              style={{
+                height: '100vh',
+                display: 'flex',
+                flexDirection: 'column',
+              }}
+            >
+              <Header>
+                <HeaderBar />
+              </Header>
+              <Layout style={{ flex: 1, overflow: 'hidden' }}>
+                <Sider>
+                  <SiderBar />
+                </Sider>
+                <Layout>
+                  <Content style={{ overflowY: 'auto', padding: '24px' }}>
+                    <App />
+                  </Content>
+                  <Layout.Footer>
+                    <FooterBar></FooterBar>
+                  </Layout.Footer>
+                </Layout>
               </Layout>
               <ToastContainer />
             </Layout>
