@@ -37,6 +37,11 @@ const STATUS_CODE_MAPPING_EXAMPLE = {
   400: '500',
 };
 
+const REGION_EXAMPLE = {
+  default: 'us-central1',
+  'claude-3-5-sonnet-20240620': 'europe-west1',
+};
+
 const fetchButtonTips =
   '1. 新建渠道时，请求通过当前浏览器发出；2. 编辑已有渠道，请求通过后端服务器发出';
 
@@ -584,6 +589,44 @@ const EditChannel = (props) => {
               />
             </>
           )}
+          {inputs.type === 41 && (
+            <>
+              <div style={{ marginTop: 10 }}>
+                <Typography.Text strong>部署地区：</Typography.Text>
+              </div>
+              <TextArea
+                name='other'
+                placeholder={
+                  '请输入部署地区，例如：us-central1\n支持使用模型映射格式\n' +
+                  '{\n' +
+                  '    "default": "us-central1",\n' +
+                  '    "claude-3-5-sonnet-20240620": "europe-west1"\n' +
+                  '}'
+                }
+                autosize={{ minRows: 2 }}
+                onChange={(value) => {
+                  handleInputChange('other', value);
+                }}
+                value={inputs.other}
+                autoComplete='new-password'
+              />
+              <Typography.Text
+                style={{
+                  color: 'rgba(var(--semi-blue-5), 1)',
+                  userSelect: 'none',
+                  cursor: 'pointer',
+                }}
+                onClick={() => {
+                  handleInputChange(
+                    'other',
+                    JSON.stringify(REGION_EXAMPLE, null, 2),
+                  );
+                }}
+              >
+                填入模板
+              </Typography.Text>
+            </>
+          )}
           {inputs.type === 21 && (
             <>
               <div style={{ marginTop: 10 }}>
@@ -730,17 +773,48 @@ const EditChannel = (props) => {
               autoComplete='new-password'
             />
           ) : (
-            <Input
-              label='密钥'
-              name='key'
-              required
-              placeholder={type2secretPrompt(inputs.type)}
-              onChange={(value) => {
-                handleInputChange('key', value);
-              }}
-              value={inputs.key}
-              autoComplete='new-password'
-            />
+            <>
+              {inputs.type === 41 ? (
+                <TextArea
+                  label='鉴权json'
+                  name='key'
+                  required
+                  placeholder={
+                    '{\n' +
+                    '  "type": "service_account",\n' +
+                    '  "project_id": "abc-bcd-123-456",\n' +
+                    '  "private_key_id": "123xxxxx456",\n' +
+                    '  "private_key": "-----BEGIN PRIVATE KEY-----xxxx\n' +
+                    '  "client_email": "xxx@developer.gserviceaccount.com",\n' +
+                    '  "client_id": "111222333",\n' +
+                    '  "auth_uri": "https://accounts.google.com/o/oauth2/auth",\n' +
+                    '  "token_uri": "https://oauth2.googleapis.com/token",\n' +
+                    '  "auth_provider_x509_cert_url": "https://www.googleapis.com/oauth2/v1/certs",\n' +
+                    '  "client_x509_cert_url": "https://xxxxx.gserviceaccount.com",\n' +
+                    '  "universe_domain": "googleapis.com"\n' +
+                    '}'
+                  }
+                  onChange={(value) => {
+                    handleInputChange('key', value);
+                  }}
+                  autosize={{ minRows: 10 }}
+                  value={inputs.key}
+                  autoComplete='new-password'
+                />
+              ) : (
+                <Input
+                  label='密钥'
+                  name='key'
+                  required
+                  placeholder={type2secretPrompt(inputs.type)}
+                  onChange={(value) => {
+                    handleInputChange('key', value);
+                  }}
+                  value={inputs.key}
+                  autoComplete='new-password'
+                />
+              )}
+            </>
           )}
           {inputs.type === 1 && (
             <>
